+++ lib/features/card_validation/domain/usecases/scan_card.dart
@@ -1,4 +1,4 @@
-import 'package:credit_card_validator/features/card_validation/domain/repositories/card_repository.dart';
+import 'package:credit_card_validator/features/card_validation/domain/repositories/repository.dart';
 
 class ScanCard {
   final CardRepository repository;
