+++ lib/features/card_validation/domain/usecases/update_banned_countries.dart
@@ -1,4 +1,4 @@
-import 'package:credit_card_validator/features/card_validation/domain/repositories/card_repository.dart';
+import 'package:credit_card_validator/features/card_validation/domain/repositories/repository.dart';
 
 class UpdateBannedCountries {
   final CardRepository repository;
