+++ lib/injection_container.dart
@@ -1,6 +1,6 @@
 import 'package:credit_card_validator/features/card_validation/data/datasources/local_data_source.dart';
-import 'package:credit_card_validator/features/card_validation/data/repositories/card_repository_impl.dart';
-import 'package:credit_card_validator/features/card_validation/domain/repositories/card_repository.dart';
+import 'package:credit_card_validator/features/card_validation/data/repositories/repository_impl.dart';
+import 'package:credit_card_validator/features/card_validation/domain/repositories/repository.dart';
 import 'package:credit_card_validator/features/card_validation/presentation/blocs/banned_countries/country_bloc.dart';
 import 'package:credit_card_validator/features/card_validation/presentation/blocs/credit_cards/card_bloc.dart';
 import 'package:get_it/get_it.dart';
